added rps cog and optimized code
